# docs/deployment/T08_handoff/acceptance_evidence/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "QuestionListResponse": {
        "properties": {
          "availability": {
            "enum": [
              "available",
              "partial",
              "unavailable"
            ],
            "title": "Availability",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "count",
          "total",
          "availability"
        ],
        "title": "QuestionListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "面向赛道 A『科学假设生成与研究计划设计』的 AI Scientist 原型 API。",
    "title": "SAGE125-AI-Scientist",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/questions": {
      "get": {
        "operationId": "list_questions_api_v1_questions_get",
        "parameters": [
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 125,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "availability": "partial",
                  "count": 1,
                  "items": [
                    {
                      "domain": "materials science",
                      "question": "How can a stable catalyst be designed?",
                      "question_id": "Q001",
                      "source_excerpt": "Design a stable catalyst.",
                      "source_page": 12,
                      "status": "unavailable",
                      "status_reason": "T07 question status was not supplied."
                    }
                  ],
                  "total": 1
                },
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponse"
                }
              }
            },
            "description": "问题列表"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INVALID_CORRELATION_ID",
                  "correlation_id": "8da1ef30-72fa-4aba-949f-4402079df58d",
                  "details": {},
                  "message": "X-Correlation-ID 格式无效。",
                  "retryable": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Correlation ID 格式无效"
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "code": "AUTHENTICATION_REQUIRED",
                  "correlation_id": "judge-demo-001",
                  "details": {},
                  "message": "需要有效的 API key。",
                  "retryable": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "需要 API key"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "FORBIDDEN",
                  "correlation_id": "judge-demo-001",
                  "details": {},
                  "message": "无权访问该资源。",
                  "retryable": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "资源不属于当前调用方"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "code": "JOB_NOT_FOUND",
                  "correlation_id": "judge-demo-001",
                  "details": {
                    "job_id": "missing-job"
                  },
                  "message": "任务不存在。",
                  "retryable": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "code": "IDEMPOTENCY_CONFLICT",
                  "correlation_id": "judge-demo-001",
                  "details": {},
                  "message": "相同 Idempotency-Key 已用于不同请求。",
                  "retryable": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "幂等冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "example": {
                  "code": "REQUEST_BODY_TOO_LARGE",
                  "correlation_id": "judge-demo-001",
                  "details": {
                    "max_bytes": 65536
                  },
                  "message": "API 请求体超过 64 KiB 上限。",
                  "retryable": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "请求体超限"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": "VALIDATION_ERROR",
                  "correlation_id": "judge-demo-001",
                  "details": {
                    "errors": []
                  },
                  "message": "请求参数不符合 API 契约。",
                  "retryable": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "请求校验失败"
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "code": "RATE_LIMIT_EXCEEDED",
                  "correlation_id": "judge-demo-001",
                  "details": {
                    "limit": 60,
                    "window_seconds": 60
                  },
                  "message": "请求频率超过限制，请稍后重试。",
                  "retryable": true
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "请求频率超限"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "code": "INTERNAL_ERROR",
                  "correlation_id": "judge-demo-001",
                  "details": {},
                  "message": "服务发生未预期错误。",
                  "retryable": false
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "未预期错误"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "code": "UPSTREAM_CONTRACT_UNAVAILABLE",
                  "correlation_id": "judge-demo-001",
                  "details": {
                    "availability": "unavailable",
                    "component": "T02/T03/T05 contract"
                  },
                  "message": "上游公开契约尚未接入。",
                  "retryable": true
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "队列或上游契约不可用"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "查询问题清单与 owner 状态",
        "tags": [
          "API v1"
        ]
      }
    },
...
    "/ingest": {
      "post": {
        "description": "接收用户上传资料并加入本地 RAG 索引（绝不外发到公开文献 API）。\n\n参数：\n    files: 上传的 PDF/TXT/MD/CSV 文件。\n\n返回：\n    {\"status\",\"files\",\"chunks_added\",\"index_status\"}；仅返回文件名。",
        "operationId": "ingest_ingest_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ingest Ingest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest"
      }
    },
...
    "/library/documents/{document_id}": {
      "delete": {
        "description": "显式删除一篇文献及其向量；系统不会定期自动删除。",
        "operationId": "delete_library_document_library_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Library Document Library Documents  Document Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Library Document"
      }
    },
...
    "/preflight": {
      "get": {
        "description": "真实模式 preflight 检查（不泄露 Key）。",
        "operationId": "preflight_real_preflight_get",
        "parameters": [
          {
            "in": "query",
            "name": "use_local_rag",
            "required": false,
            "schema": {
              "default": true,
              "title": "Use Local Rag",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "use_deep_research",
            "required": false,
            "schema": {
              "default": true,
              "title": "Use Deep Research",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preflight Real Preflight Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preflight Real"
      }
    },
...
  }
}